refactor: remove SectionalLayout as separate component & imbed this in StaticPage index file
